+++ tox.ini
@@ -8,7 +8,7 @@ ignore = W293,
 exclude = .git,
           __pycache__,
           workflows,
-          ./photosynthesis_metrics/__init__.py,
+          ./piq/__init__.py,
 
 max-complexity = 10
 max-line-length = 120
